refactor: component file structure
