+++ src/components/Popover.tsx
@@ -1,0 +1,27 @@
+import { useCallback, useState } from 'react'
+
+export type PopoverState =
+  | { isVisible: false }
+  | { isVisible: true; coordinates: { longitude: number; latitude: number }; content: JSX.Element }
+
+export function usePopoverState() {
+  const [popover, setPopover] = useState<PopoverState>({ isVisible: false })
+
+  const show = useCallback(function show(
+    coordinates: { longitude: number; latitude: number },
+    content: JSX.Element,
+  ) {
+    setPopover({ isVisible: true, coordinates, content })
+  },
+  [])
+
+  const hide = useCallback(function hide() {
+    setPopover({ isVisible: false })
+  }, [])
+
+  return {
+    ...popover,
+    show,
+    hide,
+  }
+}
