+++ src/components/hero.tsx
@@ -1,0 +1,48 @@
+export function Hero(): JSX.Element {
+  return (
+    <div className="mx-auto grid max-w-6xl place-items-center gap-8 py-48 px-8">
+      <div className="grid gap-4">
+        <h1 className="text-center text-5xl font-extrabold">Ideas Crossing the Atlantic</h1>
+        <h2 className="text-center text-2xl font-bold">
+          A Geovisualization of Transatlantic Networks and Emigration (from Central Europe)
+        </h2>
+      </div>
+      <p className="text-lg leading-relaxed">
+        During the last ten years the members of the interdisciplinary commission the North Atlantic
+        Triangle have explored the close transatlantic links between Europeans and North Americans
+        and the exchange of people and ideas in the 19th, 20th and 21st centuries. They have in
+        particular documented the networks established especially in the years between the two World
+        Wars in a number of volumes, and have underlined the importance of these relationships for
+        the refugees from Nazi Germany, including Austria after the Anschluss; who sought a safe
+        haven in the USA. In the initial phase of establishing the database the ties of friendship
+        between several of the many American foreign correspondents simultaneously stationed in
+        Vienna (!) and a number of visiting writers (as well as numerous physicians attending
+        courses arranged by the American Medical Association in Vienna) on the one hand, and members
+        of the Austrian social and cultural elite, on the other, are highlighted, and the fact that
+        the proximity of their residences in Vienna facilitated contacts and interaction is
+        demonstrated. These networks later – through affidavits supplied by American friends ¬
+        facilitated the search for safety in North America, where the refugees found support and
+        employment, especially close to the Atlantic and the Pacific coasts – particularly in
+        academic fields, and in the realms of music and the theater. This fact is visualized with
+        the help of the Geographic Information System. The database will also illustrate the
+        fortunes and activities of the (relatively few) returnees among the emigrants, and show the
+        effect and significance of the exchange of ideas across the North Atlantic, all so far
+        documented in the ten volumes published by the commission since 2012.
+      </p>
+      <p className="text-lg leading-relaxed">
+        Search in the database will eventually permit interested users to identify key agents in the
+        transatlantic cultural exchange, especially between the 1920s and 1960s. It will allow them
+        to recognize the role of encounters in the biographies of scores of individuals in Central
+        Europe and in the USA and Canada whose autobiographies and correspondence reflect these
+        ties, and permit the users to note the stimulation of professional careers in the media, the
+        arts and in academe by such contacts. It will also encourage the users to trace the impact
+        of transatlantic interaction on the genesis of non-fictional and fictional texts, on plays
+        and musical compositions, and on the evolution of (shared) philosophical and political
+        ideas. It will also help to assess the consequences of such exchanges for the perception of
+        the different societies and the development of their various institutions. The database will
+        aid the users in their attempts to comprehend the historical shifts in the relationship
+        between the countries of Europe and North America.
+      </p>
+    </div>
+  )
+}
